Feature/back/diary

See merge request s07-final/S07P31D109!109

+++ BE/climbingBear/src/main/java/com/example/climbingBear/domain/mntn/dto/MntnListResDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 public class MntnListResDto {
     private Long mntnSeq;
     private String mntnNm;
-    private Float mntnLat;
-    private Float mntnLon;
+    private double mntnLat;
+    private double mntnLon;
 
 
     public MntnListResDto(Mountain mntn){
